refactor LessonList in course view lessons
