+++ src/pages/admin/courses/view-lesson/index.js
@@ -8,7 +8,7 @@ import LessonComments from 'components/Course/LessonComments'
 import LessonMainContent from 'components/Course/LessonMainContent'
 import CourseProgressCard from 'components/Course/ProgressCard'
 import AdditionalContentCard from 'components/Course/AdditionalContentCard'
-import LessonList from 'components/Course/LessonList'
+import LessonList from 'components/Course/LessonPlaylist'
 import { Button, Space } from 'antd'
 import { CheckOutlined, CloseOutlined } from '@ant-design/icons'
 import { acceptCourseRequest, rejectCourseRequest } from 'services/courses/requests'
